+++ src/Concerns/ManagesSessionPool.php
@@ -17,7 +17,7 @@
 
 namespace Colopl\Spanner\Concerns;
 
-use Colopl\Spanner\Session;
+use Colopl\Spanner\SessionInfo;
 use Google\Auth\FetchAuthTokenInterface;
 use Google\Cloud\Spanner\Connection\ConnectionInterface;
 use Google\Cloud\Spanner\Connection\Grpc;
@@ -72,7 +72,7 @@ trait ManagesSessionPool
     }
 
     /**
-     * @return Collection<int, Session>
+     * @return Collection<int, SessionInfo>
      */
     public function listSessions(): Collection
     {
@@ -80,7 +80,7 @@ trait ManagesSessionPool
         $response = (new ProtobufSpannerClient())->listSessions($databaseName);
         return collect($response->iterateAllElements())->map(function ($session) {
             assert($session instanceof ProtobufSpannerSession);
-            return new Session($session);
+            return new SessionInfo($session);
         });
     }